Refactor frontend logic out of top.py
Move all the vast majority of frontend
logic, including anything language-specific,
to a new file called frontend.py.

Also, fix a (presumed) bug where smack_headers()
uses a missing "args" parameter. I have no idea
how this wasn't caught in daily SMACK usage.
