+++ share/smack/top.py
@@ -9,55 +9,12 @@ import shutil
 import sys
 import shlex
 import subprocess
-from svcomp.utils import svcomp_frontend
 from svcomp.utils import verify_bpl_svcomp
 from utils import temporary_file, try_command, remove_temp_files
 from replay import replay_error_trace
+from frontend import link_bc_files, frontends, languages, extra_libs 
 
 VERSION = '1.9.0'
-
-def languages():
-  """A dictionary of languages per file extension."""
-  return {
-    'c'      : 'C',
-    'i'      : 'C',
-    'cc'     : 'CXX',
-    'cpp'    : 'CXX',
-    'm'      : 'OBJC',
-    'd'      : 'D',
-    'json'   : 'JSON',
-    'svcomp' : 'SVCOMP',
-    'bc'     : 'LLVM',
-    'll'     : 'LLVM',
-    'bpl'    : 'BOOGIE',
-    'f'      : 'FORTRAN',
-    'for'    : 'FORTRAN',
-    'f90'    : 'FORTRAN',
-    'f95'    : 'FORTRAN',
-    'f03'    : 'FORTRAN',
-  }
-
-def frontends():
-  """A dictionary of front-ends per language."""
-  return {
-    'C'        : clang_frontend,
-    'CXX'      : clang_plusplus_frontend,
-    'OBJC'     : clang_objc_frontend,
-    'D'        : d_frontend,
-    'JSON'     : json_compilation_database_frontend,
-    'SVCOMP'   : svcomp_frontend,
-    'LLVM'     : llvm_frontend,
-    'BOOGIE'   : boogie_frontend,
-    'FORTRAN'  : fortran_frontend,
-  }
-
-def extra_libs():
-  """A dictionary of extra SMACK libraries required by languages."""
-  return {
-    'FORTRAN' : fortran_build_libs, 
-    'CXX'     : cplusplus_build_libs,
-    # coming soon - libraries for OBJC, Rust, Swift, etc.
-  }
 
 def results(args):
   """A dictionary of the result output messages."""
@@ -330,225 +287,6 @@ def frontend(args):
   
   if not noreturning_frontend:
     return link_bc_files(bitcodes,libs,args)
-
-def smack_root():
-  return os.path.dirname(os.path.dirname(os.path.abspath(sys.argv[0])))
-
-def smack_header_path():
-  return os.path.join(smack_root(), 'share', 'smack', 'include')
-
-def smack_headers():
-  paths = []
-  paths.append(smack_header_path())
-  if args.memory_safety or args.integer_overflow:
-    paths.append(os.path.join(smack_header_path(), 'string'))
-  if args.float:
-    paths.append(os.path.join(smack_header_path(), 'math'))
-  return paths
-
-def smack_lib():
-  return os.path.join(smack_root(), 'share', 'smack', 'lib')
-
-def default_clang_compile_command(args, lib = False):
-  cmd = ['clang', '-c', '-emit-llvm', '-O0', '-g', '-gcolumn-info']
-  cmd += map(lambda path: '-I' + path, smack_headers())
-  cmd += args.clang_options.split()
-  cmd += ['-DMEMORY_MODEL_' + args.mem_mod.upper().replace('-','_')]
-  if args.memory_safety: cmd += ['-DMEMORY_SAFETY']
-  if args.integer_overflow: cmd += (['-ftrapv'] if not lib else ['-DSIGNED_INTEGER_OVERFLOW_CHECK'])
-  if args.float: cmd += ['-DFLOAT_ENABLED']
-  if sys.stdout.isatty(): cmd += ['-fcolor-diagnostics']
-  return cmd
-
-def default_build_libs(args):
-  """Generate LLVM bitcodes for SMACK libraries."""
-  bitcodes = []
-  libs = ['smack.c']
-
-  if args.pthread:
-    libs += ['pthread.c']
-
-  if args.strings or args.memory_safety or args.integer_overflow:
-    libs += ['string.c']
-
-  if args.float:
-    libs += ['math.c']
-    libs += ['fenv.c']
-
-  compile_command = default_clang_compile_command(args, True)
-  for c in map(lambda c: os.path.join(smack_lib(), c), libs):
-    bc = compile_to_bc(c,compile_command,args)
-    bitcodes.append(bc)
-
-  return bitcodes
-
-def fortran_build_libs(args):
-  """Generate FORTRAN-specific LLVM bitcodes for SMACK libraries."""
-
-  bitcodes = []
-  libs = ['smack.f90']
-
-  compile_command = default_clang_compile_command(args)
-  compile_command[0] = 'flang'
-
-  for c in map(lambda c: os.path.join(smack_lib(), c), libs):
-    bc = fortran_compile_to_bc(c,compile_command,args)
-    bitcodes.append(bc)
-
-  return bitcodes
-
-def cplusplus_build_libs(args):
-  """Generate C++ specific LLVM bitcodes for SMACK libraries."""
-
-  bitcodes = []
-  libs = ['smack.cpp']
-
-  compile_command = default_clang_compile_command(args,True)
-  compile_command[0] = 'clang++'
-
-  for c in map(lambda c: os.path.join(smack_lib(), c), libs):
-    bc = compile_to_bc(c,compile_command,args)
-    bitcodes.append(bc)
-
-  return bitcodes
-
-
-def llvm_frontend(input_file, args):
-  """Return LLVM IR file. Exists for symmetry with other frontends."""
-
-  return input_file
-
-def clang_frontend(input_file, args):
-  """Generate LLVM IR from C-language source(s)."""
-
-  compile_command = default_clang_compile_command(args)
-  return compile_to_bc(input_file,compile_command,args)
-
-def clang_plusplus_frontend(input_file, args):
-  """Generate LLVM IR from C++ language source(s)."""
-  compile_command = default_clang_compile_command(args)
-  compile_command[0] = 'clang++'
-  return compile_to_bc(input_file,compile_command,args)
-
-def clang_objc_frontend(input_file, args):
-  """Generate LLVM IR from Objective-C language source(s)."""
-
-  compile_command = default_clang_compile_command(args)
-  if sys.platform in ['linux', 'linux2']:
-    objc_flags = try_command(['gnustep-config', '--objc-flags'])
-    compile_command += objc_flags.split()
-  elif sys.platform == 'darwin':
-    sys.exit("Objective-C not yet supported on macOS")
-  else:
-    sys.exit("Objective-C not supported for this operating system.")
-  return compile_to_bc(input_file,compile_command,args)
-
-def d_frontend(input_file, args):
-  """Generate Boogie code from D programming language source(s)."""
-
-  # note: -g and -O0 are not used here. 
-  # Right now, it works, and with these options, smack crashes.
-  compile_command = ['ldc2', '-output-ll'] 
-  compile_command += map(lambda path: '-I=' + path, smack_headers())
-  args.entry_points += ['_Dmain']
-  return compile_to_bc(input_file,compile_command,args)
-
-def fortran_frontend(input_file, args):
-  """Generate Boogie code from Fortran language source(s)."""
-
-  #  For a fortran file that includes smack.f90 as a module,
-  #  it will not compile unless the file 'smack.mod' exists
-  #  in the working directory. 'smack.mod' is a build artifact
-  #  of compiling smack.f90. Therefore, the solution is to 
-  #  compile smack.f90 before the source files.
-  fortran_build_libs(args)
-  #  The result of this computation will be discarded when SMACK
-  #  builds it's libraries later.
-
-  # replace the default entry point with the fortran default 'MAIN_'
-  args.entry_points += ['MAIN_']
-
-  compile_command = default_clang_compile_command(args)
-  compile_command[0] = 'flang'
-
-  return fortran_compile_to_bc(input_file,compile_command,args)
-
-def boogie_frontend(input_file, args):
-  """Pass Boogie code to the verifier."""
-  if len(args.input_files) > 1: 
-      raise RuntimeError("Expected a single Boogie file.")
-
-  with open(args.bpl_file, 'a+') as out:
-    with open(input_file) as f:
-      out.write(f.read())
-
-def json_compilation_database_frontend(input_file, args):
-  """Generate Boogie code from a JSON compilation database."""
-
-  if len(args.input_files) > 1:
-    raise RuntimeError("Expected a single JSON compilation database.")
-
-  output_flags = re.compile(r"-o ([^ ]*)[.]o\b")
-  optimization_flags = re.compile(r"-O[1-9]\b")
-
-  with open(input_file) as f:
-    for cc in json.load(f):
-      if 'objects' in cc:
-        # TODO what to do when there are multiple linkings?
-        bit_codes = map(lambda f: re.sub('[.]o$','.bc',f), cc['objects'])
-        try_command(['llvm-link', '-o', args.bc_file] + bit_codes)
-        try_command(['llvm-link', '-o', args.linked_bc_file, args.bc_file] + build_libs(args))
-
-      else:
-        out_file = output_flags.findall(cc['command'])[0] + '.bc'
-        command = cc['command']
-        command = output_flags.sub(r"-o \1.bc", command)
-        command = optimization_flags.sub("-O0", command)
-        command = command + " -emit-llvm"
-        try_command(command.split(),cc['directory'], console=True)
-
-  llvm_to_bpl(args)
-
-def compile_to_bc(input_file, compile_command, args):
-  """Compile a source file to LLVM IR."""
-  bc = temporary_file(os.path.splitext(os.path.basename(input_file))[0], '.bc', args)
-  try_command(compile_command + ['-o', bc, input_file], console=True)
-  return bc
-
-def fortran_compile_to_bc(input_file, compile_command, args):
-  """Compile a FORTRAN source file to LLVM IR."""
-
-  #  This method only exists as a hack to get flang to work
-  #  with SMACK. When we update to the latest flang on LLVM 5,
-  #  this method will no longer be necessary. The hack is 
-  #  self-contained in this method.
-
-  #  The Debug Info Version in flang is incompatible with 
-  #  the version that clang uses. The workaround is to use
-  #  sed to change the file so llvm-link gives a warning
-  #  and not an error. 
-
-  # compile to human-readable format in order to tweak the IR
-  compile_command[1] = '-S'
-  ll = temporary_file(os.path.splitext(os.path.basename(input_file))[0], '.ll', args)
-  try_command(compile_command + ['-o', ll, input_file], console=True)
-  # change the throw level of 'Debug Info Version' from error to warning in the IR
-  try_command(['sed', '-i', 's/i32 1, !\"Debug Info Version\"/i32 2, !\"Debug Info Version\"/g', ll])
-  try_command(['llvm-as', ll])
-  try_command(['rm', ll])
-  bc = '.'.join(ll.split('.')[:-1] + ['bc'])
-  return bc 
- 
-def link_bc_files(bitcodes, libs, args):
-  """Link generated LLVM bitcode and relevant smack libraries."""
-
-  smack_libs = default_build_libs(args)
-  for build_lib in libs:
-    smack_libs += build_lib(args)
-
-  try_command(['llvm-link', '-o', args.bc_file] + bitcodes)
-  try_command(['llvm-link', '-o', args.linked_bc_file, args.bc_file] + smack_libs)
-  llvm_to_bpl(args)
 
 def llvm_to_bpl(args):
   """Translate the LLVM bitcode file to a Boogie source file."""
